requirements.txt
python-dotenv
requests
python-telegram-bot==13.5
